Moved functions to the right header
Some functions are declared in a header but defined to a not
corresponding source file. This cause confusion and lesser the
readability.

Issue: #1
