+++ source/common.hpp
@@ -92,13 +92,3 @@ static inline color_t FetchPixelColor(uint32_t x, uint32_t y)
     u8 b = (u8)(val>>16);
     return MakeColor(r, g, b, 255);
 }
-
-void DrawPixel(uint32_t x, uint32_t y, color_t clr);
-struct coord DrawText(u32 font, uint32_t x, uint32_t y, color_t clr, const char* text);
-struct coord DrawTextTruncateW(u32 font, uint32_t x, uint32_t y, color_t clr, const char* text, uint32_t max_width, const char* end_text);
-int DrawTextTruncateH(u32 font, uint32_t x, uint32_t y, color_t clr, const char *text, int start_line, uint32_t max_height, const char *end_text);
-void GetTextDimensions(u32 font, const char* text, uint32_t* width_out, uint32_t* height_out);
-std::string WrapText(u32 font, const char *text, uint32_t max_width);
-
-bool fontInitialize(void);
-void fontExit();
